Complete Project
TicTacToe Game containing a single player mode against CPU player using the minimax algorithm, and an option to play multiplayer (locally)
This project is fairly small and not completely flawless, and I'm sure It can be done in a more optimal way, but It was a good practice to the minimax algorithm and most importantly, It was pretty fun! I will stop working on it here and move on to bigger projects, thank you for everyone who'll check it out!